FlatPostgresGroupDAO
