+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/directory/FlatPostgresDirectoryDAO.scala
@@ -6,7 +6,7 @@ import org.broadinstitute.dsde.workbench.sam.errorReportSource
 import org.broadinstitute.dsde.workbench.model.{ErrorReport, WorkbenchEmail, WorkbenchException, WorkbenchExceptionWithErrorReport, WorkbenchGroupIdentity, WorkbenchGroupName, WorkbenchSubject, WorkbenchUserId}
 import org.broadinstitute.dsde.workbench.sam.db.SamParameterBinderFactory.SqlInterpolationWithSamBinders
 import org.broadinstitute.dsde.workbench.sam.db._
-import org.broadinstitute.dsde.workbench.sam.db.tables.{FlatGroupMemberRecord, FlatGroupMemberTable, FlatGroupMembershipPath, GroupPK, GroupRecord, GroupTable, PolicyTable, ResourceTable, ResourceTypeTable}
+import org.broadinstitute.dsde.workbench.sam.db.tables.{FlatGroupMemberRecord, FlatGroupMemberTable, GroupPK, GroupRecord, GroupTable, PolicyTable, ResourceTable, ResourceTypeTable}
 import org.broadinstitute.dsde.workbench.sam.model.{AccessPolicyName, BasicWorkbenchGroup, FullyQualifiedPolicyId, FullyQualifiedResourceId, ResourceId, ResourceTypeName}
 import org.broadinstitute.dsde.workbench.sam.util.SamRequestContext
 import org.postgresql.util.PSQLException
@@ -37,58 +37,6 @@ class FlatPostgresDirectoryDAO (override val dbRef: DbReference, override val ec
     runInTransaction("addGroupMember", samRequestContext)({ implicit session =>
       insertGroupMembers(queryForGroupPKs(Set(groupId)).head, Set(addMember))
     }).map(count => count > 0)
-  }
-
-  override def insertGroupMembers(groupId: GroupPK, members: Set[WorkbenchSubject])(implicit session: DBSession): Int = {
-    if (members.isEmpty) {
-      0
-    } else {
-      val userMembers: Set[WorkbenchUserId] = members.collect { case userId: WorkbenchUserId => userId }
-      val groupMembers = queryForGroupPKs(members);
-
-      // direct (non-transitive) membership in groupId
-      val directMembershipPath = FlatGroupMembershipPath(List(groupId));
-
-      val directMemberUsers: List[SQLSyntax] = userMembers.map { userId => {
-        samsqls"(${groupId}, ${userId}, ${None}, ${directMembershipPath})"
-      }
-      }.toList
-
-      val directMemberGroups: List[SQLSyntax] = groupMembers.map { groupPK =>
-        samsqls"(${groupId}, ${None}, ${groupPK}, ${directMembershipPath})"
-      }
-
-      // groups which have groupId as a direct or transitive subgroup
-      val ancestorMemberships = listMyGroupRecords(groupId)
-
-      // groups and users which have `groupMembers` as ancestors
-      val descendantMemberships = listMembersByPKs(groupMembers)
-
-      val transitiveMembers = ancestorMemberships flatMap { case FlatGroupMemberRecord(_, ancestor, _, _, ancestorPath) =>
-        val ancestorsPlusGroup = ancestorPath.append(groupId)
-
-        val ancestorUsers = userMembers.map { userId =>
-          samsqls"(${ancestor}, ${userId}, ${None}, ${ancestorsPlusGroup})"
-        }.toList
-
-        val ancestorGroups = groupMembers.map { groupPK =>
-          samsqls"(${ancestor}, ${None}, ${groupPK}, ${ancestorsPlusGroup})"
-        }
-
-        val descendants = descendantMemberships.collect {
-          case FlatGroupMemberRecord(_, _, Some(memberUserId), _, groupMembershipPath) =>
-            samsqls"(${ancestor}, ${memberUserId}, ${None}, ${ancestorsPlusGroup.append(groupMembershipPath)})"
-          case FlatGroupMemberRecord(_, _, _, Some(memberGroupId), groupMembershipPath) =>
-            samsqls"(${ancestor}, ${None}, ${memberGroupId}, ${ancestorsPlusGroup.append(groupMembershipPath)})"
-        }
-
-        ancestorUsers ++ ancestorGroups ++ descendants
-      }
-
-      val gm = FlatGroupMemberTable.column
-      samsql"""insert into ${FlatGroupMemberTable.table} (${gm.groupId}, ${gm.memberUserId}, ${gm.memberGroupId}, ${gm.groupMembershipPath})
-        values ${directMemberUsers ++ directMemberGroups ++ transitiveMembers}""".update().apply()
-    }
   }
 
   override def removeGroupMember(groupId: WorkbenchGroupIdentity, removeMember: WorkbenchSubject, samRequestContext: SamRequestContext): IO[Boolean] = {
@@ -278,10 +226,6 @@ class FlatPostgresDirectoryDAO (override val dbRef: DbReference, override val ec
     samsqls"${f.groupMembershipPath}[array_upper(${f.groupMembershipPath}, 1)] = ${workbenchGroupIdentityToGroupPK(groupId)}"
   }
 
-  // there is probably some implicit magic which avoids this but I don't know it
-  private def convertToFlatGroupMemberTable(f: QuerySQLSyntaxProvider[SQLSyntaxSupport[FlatGroupMemberRecord], FlatGroupMemberRecord])
-                                           (rs: WrappedResultSet): FlatGroupMemberRecord = FlatGroupMemberTable.apply(f)(rs)
-
   // adaptation of PostgresDirectoryDAO.listMemberOfGroups
   private def listMemberOfGroups(subject: WorkbenchSubject, samRequestContext: SamRequestContext): IO[Set[WorkbenchGroupIdentity]]  = {
     val f = FlatGroupMemberTable.syntax("f")
@@ -309,22 +253,6 @@ class FlatPostgresDirectoryDAO (override val dbRef: DbReference, override val ec
     })
   }
 
-  // get all of the groups that group `member` is a member of - directly or transitively
-  private def listMyGroupRecords(member: GroupPK)(implicit session: DBSession) = {
-    val gm = FlatGroupMemberTable.column
-    val f = FlatGroupMemberTable.syntax("f")
-    val query = samsql"select * from ${FlatGroupMemberTable as f} where ${gm.memberGroupId} = ${member}"
-    query.map(convertToFlatGroupMemberTable(f)).list.apply()
-  }
-
-  // get all of the direct or transitive members of all `groups`
-  private def listMembersByPKs(groups: Traversable[GroupPK])(implicit session: DBSession) = {
-    val gm = FlatGroupMemberTable.column
-    val f = FlatGroupMemberTable.syntax("f")
-    val query = samsql"select * from ${FlatGroupMemberTable as f} where ${gm.groupId} IN (${groups})"
-    query.map(convertToFlatGroupMemberTable(f)).list.apply()
-  }
-
   // get all of the direct or transitive members of all `groups`
   private def listMembersByGroupIdentity(groups: Traversable[WorkbenchGroupIdentity])(implicit session: DBSession) = {
     val gm = FlatGroupMemberTable.column
@@ -332,4 +260,9 @@ class FlatPostgresDirectoryDAO (override val dbRef: DbReference, override val ec
     val query = samsql"select * from ${FlatGroupMemberTable as f} where ${gm.groupId} IN (${groups map workbenchGroupIdentityToGroupPK})"
     query.map(convertToFlatGroupMemberTable(f)).list.apply()
   }
+
+  // there is probably some implicit magic which avoids this but I don't know it
+  private def convertToFlatGroupMemberTable(f: QuerySQLSyntaxProvider[SQLSyntaxSupport[FlatGroupMemberRecord], FlatGroupMemberRecord])
+                                           (rs: WrappedResultSet): FlatGroupMemberRecord = FlatGroupMemberTable.apply(f)(rs)
+
 }
